requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
fastapi
python-multipart
python-dotenv
uvicorn
unstructured
markdown
